+++ src/Codegen.hs
@@ -9,28 +9,30 @@ import qualified Client.Game
 import qualified Client.Room
 import qualified Game
 import Servant.Elm (DefineElm (DefineElm), Proxy (Proxy), defElmImports, defElmOptions, generateElmModuleWith)
-import Server (APIRoutes)
+-- import Server (APIRoutes)
 import System.Process (createProcess, cwd, proc)
 
 run :: IO ()
 run =
-  let outputDir = "./ui/generated"
-   in do
-        putStrLn "Deleting old generated code..."
-        putStrLn "Generating Elm code from API..."
-        generateElmModuleWith
-          defElmOptions
-          ["Api"]
-          defElmImports
-          outputDir
-          [ DefineElm (Proxy :: Proxy Game.Country),
-            DefineElm (Proxy :: Proxy Game.Color),
-            DefineElm (Proxy :: Proxy Client.Game.Game),
-            DefineElm (Proxy :: Proxy Client.Game.Instructions),
-            DefineElm (Proxy :: Proxy Client.ConnectionStates.ConnectionStates),
-            DefineElm (Proxy :: Proxy Client.Room.Room)
-          ]
-          (Proxy :: Proxy APIRoutes)
-        putStrLn "Formatting generated code using elm-format..."
-        createProcess (proc "elm-format" ["Api.elm", "--yes"]) {cwd = Just outputDir}
-        putStrLn "Done!"
+  return ()
+
+-- let outputDir = "./ui/generated"
+--  in do
+--       putStrLn "Deleting old generated code..."
+--       putStrLn "Generating Elm code from API..."
+--       generateElmModuleWith
+--         defElmOptions
+--         ["Api"]
+--         defElmImports
+--         outputDir
+--         [ DefineElm (Proxy :: Proxy Game.Country),
+--           DefineElm (Proxy :: Proxy Game.Color),
+--           DefineElm (Proxy :: Proxy Client.Game.Game),
+--           DefineElm (Proxy :: Proxy Client.Game.Instructions),
+--           DefineElm (Proxy :: Proxy Client.ConnectionStates.ConnectionStates),
+--           DefineElm (Proxy :: Proxy Client.Room.Room)
+--         ]
+--         (Proxy :: Proxy APIRoutes)
+--       putStrLn "Formatting generated code using elm-format..."
+--       createProcess (proc "elm-format" ["Api.elm", "--yes"]) {cwd = Just outputDir}
+--       putStrLn "Done!"
